+++ client/src/views/old/Login.vue
@@ -2,7 +2,7 @@
   <login-card></login-card>
 </template>
 <script>
-import LoginCard from "@/components/LoginCard";
+import LoginCard from "@/components/old/LoginCard";
 export default {
   components: {LoginCard}
 }
